Some of my recent work

+++ PoGo.NecroBot.Logic/Tasks/TransferPokemonTask.cs
@@ -13,39 +13,44 @@ namespace PoGo.NecroBot.Logic.Tasks
     {
         public static async Task Execute(ISession session, ulong pokemonId)
         {
-            var all = await session.Inventory.GetPokemons();
-            var pokemons = all.OrderBy(x => x.Cp).ThenBy(n => n.StaminaMax);
-            var pokemon = pokemons.FirstOrDefault(p => p.Id == pokemonId);
-
-            if (pokemon == null) return;
-
-            var pokemonSettings = await session.Inventory.GetPokemonSettings();
-            var pokemonFamilies = await session.Inventory.GetPokemonFamilies();
-
-            await session.Client.Inventory.TransferPokemon(pokemonId);
-            await session.Inventory.DeletePokemonFromInvById(pokemonId);
-
-            var bestPokemonOfType = (session.LogicSettings.PrioritizeIvOverCp
-                ? await session.Inventory.GetHighestPokemonOfTypeByIv(pokemon)
-                : await session.Inventory.GetHighestPokemonOfTypeByCp(pokemon)) ?? pokemon;
-
-            var setting = pokemonSettings.Single(q => q.PokemonId == pokemon.PokemonId);
-            var family = pokemonFamilies.First(q => q.FamilyId == setting.FamilyId);
-
-            family.Candy_++;
-
-            // Broadcast event as everyone would benefit
-            session.EventDispatcher.Send(new Logic.Event.TransferPokemonEvent
+            using (var blocker = new BlockableScope(session, Model.BotActions.Transfer))
             {
-                Id = pokemon.PokemonId,
-                Perfection = Logic.PoGoUtils.PokemonInfo.CalculatePokemonPerfection(pokemon),
-                Cp = pokemon.Cp,
-                BestCp = bestPokemonOfType.Cp,
-                BestPerfection = Logic.PoGoUtils.PokemonInfo.CalculatePokemonPerfection(bestPokemonOfType),
-                FamilyCandies = family.Candy_
-            });
+                if (!await blocker.WaitToRun()) return;
 
-            DelayingUtils.Delay(session.LogicSettings.TransferActionDelay, 0);
+                var all = await session.Inventory.GetPokemons();
+                var pokemons = all.OrderBy(x => x.Cp).ThenBy(n => n.StaminaMax);
+                var pokemon = pokemons.FirstOrDefault(p => p.Id == pokemonId);
+
+                if (pokemon == null) return;
+
+                var pokemonSettings = await session.Inventory.GetPokemonSettings();
+                var pokemonFamilies = await session.Inventory.GetPokemonFamilies();
+
+                await session.Client.Inventory.TransferPokemon(pokemonId);
+                await session.Inventory.DeletePokemonFromInvById(pokemonId);
+
+                var bestPokemonOfType = (session.LogicSettings.PrioritizeIvOverCp
+                    ? await session.Inventory.GetHighestPokemonOfTypeByIv(pokemon)
+                    : await session.Inventory.GetHighestPokemonOfTypeByCp(pokemon)) ?? pokemon;
+
+                var setting = pokemonSettings.Single(q => q.PokemonId == pokemon.PokemonId);
+                var family = pokemonFamilies.First(q => q.FamilyId == setting.FamilyId);
+
+                family.Candy_++;
+
+                // Broadcast event as everyone would benefit
+                session.EventDispatcher.Send(new Logic.Event.TransferPokemonEvent
+                {
+                    Id = pokemon.PokemonId,
+                    Perfection = Logic.PoGoUtils.PokemonInfo.CalculatePokemonPerfection(pokemon),
+                    Cp = pokemon.Cp,
+                    BestCp = bestPokemonOfType.Cp,
+                    BestPerfection = Logic.PoGoUtils.PokemonInfo.CalculatePokemonPerfection(bestPokemonOfType),
+                    FamilyCandies = family.Candy_
+                });
+
+                DelayingUtils.Delay(session.LogicSettings.TransferActionDelay, 0);
+            }
         }
     }
 }
